refactor: the NMI sources live with the CPU
